refactor(vite-plugin-angular): move shared utils to utils/shared.ts
Extract DiagnosticModes, injectViteIgnoreForHmrMetadata, isIgnoredHmrFile,
toAngularCompilationFileReplacements, mapTemplateUpdatesToFiles,
refreshStylesheetRegistryForFile, and describeStylesheetContent into
utils/shared.ts so both angular-vite-plugin.ts and compilation-api-plugin.ts
import from a shared location instead of compilation-api depending on
the main plugin file.
